requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Flask==3.0.2
fonttools==4.49.0
impedance==1.7.1
importlib_metadata==7.0.2
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
pyparsing==3.1.2
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.33.0
rpds-py==0.18.0
scipy==1.10.1
six==1.16.0
toolz==0.12.1
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1
zipp==3.17.0
